modify: jobinfo api
