+++ backend/server.js
@@ -24,7 +24,7 @@ app.use(jobsRoutes);
 const applicationRoutes = require("./application/route");
 app.use(applicationRoutes);
 
-const landlordRoutes = require("./landlord/route");
+const landlordRoutes = require("./jobs/route");
 app.use(landlordRoutes);
 
 app.listen(PORT, () => {
